refactor(slime): share one LoRA request schema check
preflight.py and sglang/engine.py each carried their own copy of the same two
checks, differing only in function name and a trailing hint in the error. Both
copies spell out the required field sets, so an SGLang upgrade that renames a
field has to be tracked in two places, and missing one turns a startup error
into a runtime failure. The test file guarded the duplication by asserting the
two copies behave identically.

The field sets and both checks now live in sglang/lora_schema.py, keeping the
message that names Reef's pinned image because the driver preflight benefits
from it as much as the engine actor. The checks import SGLang inside the
function, as before, so the driver still reaches them without a GPU stack.
